Commit original : th202608  2014-01-06 14:23:53 +0100 (lun. 06 janv. 2014)
add the ability to select multiple finite strain strategies in the UMAT interface. Associated tests. Add a test on the exponential of a tensor and its derivatives based on an article of Miehe. Correct the computation of the second piola-kirchhoff stress from the logarithmic stress (correct computation of the derivative of the logaritm of a tensor when two or more eigenvalues are equals).